configure a secure web connection
