+++ commonvoice/api/app.py
@@ -10,7 +10,7 @@ from flask_socketio import SocketIO
 
 from audio_model.audio_model.config.config import Gender, FRAME
 from audio_model.audio_model.pipeline_mananger import load_model
-from audio_model.audio_model.utils import audio_melspectrogram, generate_pred
+from audio_model.audio_model.utils import audio_melspectrogram, generate_pred, remove_silence
 
 warnings.filterwarnings("ignore")
 
@@ -64,6 +64,7 @@ def run_audio_stream(msg):
             socketio.sleep(0.5)
 
             signal = np.concatenate(tuple(frames))
+            signal = remove_silence(signal)
             wave_period = signal[-FRAME["SAMPLE_RATE"]:].astype(np.float)
             spectrogram = audio_melspectrogram(wave_period)
