--- server/default.conf
@@ -1,11 +1,0 @@
-server {
-    listen 80;
-    server_name commmvoice.me www.commmvoice.me;
-
-    location / {
-        proxy_pass http://commvoice:8080;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-        proxy_redirect off;
-    }
-
-}
