--- server/gunicorn.socket
@@ -1,14 +1,0 @@
-[Unit]
-Description=gunicorn socket
-
-[Socket]
-ListenStream=/run/gunicorn.sock
-# Our service won't need permissions for the socket, since it
-# inherits the file descriptor by socket activation
-# only the nginx daemon will need access to the socket
-SocketUser=www-data
-# Optionally restrict the socket permissions even more.
-# SocketMode=600
-
-[Install]
-WantedBy=sockets.target
