+++ server/nginxconfig.io/letsencrypt.conf
@@ -1,0 +1,4 @@
+# ACME-challenge
+location ^~ /.well-known/acme-challenge/ {
+    root /var/www/_letsencrypt;
+}
