+++ server/sites-available/commvoice.me.conf
@@ -1,0 +1,52 @@
+server {
+    listen                  443 ssl http2;
+    listen                  [::]:443 ssl http2;
+    server_name             www.commvoice.me;
+    root                    /var/www/commvoice.me/public;
+
+    # SSL
+    ssl_certificate         /etc/letsencrypt/live/commvoice.me/fullchain.pem;
+    ssl_certificate_key     /etc/letsencrypt/live/commvoice.me/privkey.pem;
+    ssl_trusted_certificate /etc/letsencrypt/live/commvoice.me/chain.pem;
+
+    # security
+    include                 nginxconfig.io/security.conf;
+
+    # index.html fallback
+    location / {
+        try_files $uri $uri/ /index.html;
+    }
+
+    # index.php fallback
+    location ~ ^/api/ {
+        try_files $uri $uri/ /index.php?$query_string;
+    }
+
+    # additional config
+    include nginxconfig.io/general.conf;
+}
+
+# non-www, subdomains redirect
+server {
+    listen                  443 ssl http2;
+    listen                  [::]:443 ssl http2;
+    server_name             .commvoice.me;
+
+    # SSL
+    ssl_certificate         /etc/letsencrypt/live/commvoice.me/fullchain.pem;
+    ssl_certificate_key     /etc/letsencrypt/live/commvoice.me/privkey.pem;
+    ssl_trusted_certificate /etc/letsencrypt/live/commvoice.me/chain.pem;
+    return                  301 https://www.commvoice.me$request_uri;
+}
+
+# HTTP redirect
+server {
+    listen      80;
+    listen      [::]:80;
+    server_name .commvoice.me;
+    include     nginxconfig.io/letsencrypt.conf;
+
+    location / {
+        return 301 https://www.commvoice.me$request_uri;
+    }
+}
